# internal/handlers/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: otus-app API
  description: |
    HTTP-сервис на Go с хранением сообщений в PostgreSQL.

    Прямые маршруты (`/messages`, `/hello`, ...) обслуживает app,
    маршруты `/gw/...` — gateway: обычное проксирование по HTTP+JSON
    и мост в gRPC (`/gw/grpc/...`) с разными видами стриминга.

    `POST /messages` и `POST /gw/grpc/...` требуют JWT из Keycloak —
    нажмите Authorize и вставьте access_token.
  version: "1.0"
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен из Keycloak (realm otus, клиент otus-app)
  schemas:
    Message:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 42
        text:
          type: string
          example: привет
        checksum:
          type: string
          description: SHA-256 текста, вычисляется при сохранении
        checksum_ok:
          type: boolean
          description: Сходится ли контрольная сумма при чтении
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
          example: требуется токен авторизации
paths:
  /health:
    get:
      summary: Проверка живости сервиса
      tags: [app]
      responses:
        "200":
          description: Сервис работает
          content:
            application/json:
              example: { "status": "работает" }
  /version:
    get:
      summary: Версия и дата сборки
      tags: [app]
      responses:
        "200":
          description: Информация о сборке
          content:
            application/json:
              example: { "version": "v1.0.0", "date": "2026-07-21T10:00:00Z" }
  /hello:
    get:
      summary: Приветствие
      tags: [app]
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          example: Женя
      responses:
        "200":
          description: Приветствие
          content:
            application/json:
              example: { "message": "Привет, Женя!" }
  /status:
    get:
      summary: Статусная страница сервиса (HTML)
      description: Отрисовывается шаблонизатором html/template, обновляется каждые 10 секунд
      tags: [app]
      responses:
        "200":
          description: HTML-страница с состоянием сервиса
          content:
            text/html: {}
  /audit:
    get:
      summary: Аудит-лог созданных сообщений
      description: Последние события из MongoDB (replica set); 503, если аудит-лог не настроен
      tags: [app]
      responses:
        "200":
          description: Последние 10 событий
          content:
            application/json:
              example:
                - { "event": "создано сообщение", "text": "привет", "at": "2026-08-14T10:00:00Z" }
        "503":
          description: Аудит-лог не настроен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /messages:
    get:
      summary: Список последних сообщений
      tags: [app]
      responses:
        "200":
          description: До 100 последних сообщений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      summary: Создать сообщение (закрытый раздел)
      description: |
        Доступ контролирует middleware — нужен JWT из Keycloak.
        Повтор запроса с тем же Idempotency-Key не создаёт дубликат,
        а возвращает уже сохранённое сообщение.
      tags: [app]
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: Ключ идемпотентности запроса
          example: 4f2a9c1e8b7d63a0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: сообщение через swagger
      responses:
        "201":
          description: Сообщение сохранено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Некорректное тело запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Нет или недействительный токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /search:
    get:
      summary: Поиск по тексту сообщений
      description: |
        Один и тот же запрос можно выполнить тремя движками и сравнить:
        полнотекстовый индекс Elasticsearch (по умолчанию), перебор
        таблицы в Cassandra и LIKE в PostgreSQL.
      tags: [storage]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          example: отчет
        - name: engine
          in: query
          required: false
          schema:
            type: string
            enum: [elasticsearch, cassandra, postgres]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Найденные сообщения и время поиска
          content:
            application/json:
              example:
                engine: elasticsearch
                query: отчет
                total: 3
                took_ms: 4
                elapsed_ms: 7
                messages: []
        "400":
          description: Не задан запрос или неизвестный движок
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Движок поиска не настроен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /events:
    get:
      summary: Лента событий из Cassandra
      description: Последние события за сутки из партиции message_events
      tags: [storage]
      parameters:
        - name: day
          in: query
          required: false
          schema:
            type: string
          example: "2026-08-31"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: События за день
          content:
            application/json:
              example:
                - { "id": 42, "text": "привет", "checksum": "9f86d0...", "producer": "kafka", "created_at": "2026-08-31T10:00:00Z" }
        "503":
          description: Cassandra не настроена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /events/{id}:
    get:
      summary: Событие по идентификатору
      description: Чтение из денормализованной таблицы events_by_id
      tags: [storage]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Событие найдено
          content:
            application/json:
              example: { "id": 42, "text": "привет", "checksum": "9f86d0...", "producer": "kafka", "created_at": "2026-08-31T10:00:00Z" }
        "404":
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files:
    get:
      summary: Список файлов и поиск по тегам
      description: |
        source=s3 перебирает бакет и читает теги каждого объекта,
        source=es отвечает по индексу Elasticsearch, source=hdfs
        показывает каталог HDFS.
      tags: [storage]
      parameters:
        - name: backend
          in: query
          required: false
          schema:
            type: string
            enum: [s3, hdfs]
        - name: tag
          in: query
          required: false
          schema:
            type: string
          example: kind=report
        - name: source
          in: query
          required: false
          schema:
            type: string
            enum: [s3, es, hdfs]
        - name: prefix
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Найденные файлы и время выборки
          content:
            application/json:
              example:
                source: elasticsearch
                total: 2
                took_ms: 3
                elapsed_ms: 6
                files: []
    post:
      summary: Загрузить файл (закрытый раздел)
      description: |
        Тело запроса — содержимое файла потоком. При неизвестном размере
        клиент S3 сам разбивает поток на части и грузит multipart.
        backend=hdfs кладет файл в HDFS с репликацией.
      tags: [storage]
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          example: report.bin
        - name: backend
          in: query
          required: false
          schema:
            type: string
            enum: [s3, hdfs]
        - name: tags
          in: query
          required: false
          schema:
            type: string
          example: kind=report,owner=evgenza
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Файл сохранен
          content:
            application/json:
              example:
                backend: s3
                object:
                  key: report.bin
                  size: 104857600
                  etag: "b1946ac9-7"
                  version_id: "b6ff7d4a"
                  parts: 7
        "400":
          description: Не задано имя файла
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Нет или недействительный токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Хранилище не приняло файл
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/{key}:
    get:
      summary: Скачать файл целиком или частично
      description: |
        Заголовок Range (bytes=начало-конец) передается в хранилище,
        ответ приходит со статусом 206 и заголовком Content-Range.
      tags: [storage]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: backend
          in: query
          required: false
          schema:
            type: string
            enum: [s3, hdfs]
        - name: version
          in: query
          required: false
          schema:
            type: string
          description: Версия объекта в S3
        - name: Range
          in: header
          required: false
          schema:
            type: string
          example: bytes=0-1023
      responses:
        "200":
          description: Файл целиком
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "206":
          description: Запрошенный диапазон байт
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Файл не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/{key}/versions:
    get:
      summary: Версии объекта в S3
      tags: [storage]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Список версий, свежая помечена is_latest
          content:
            application/json:
              example:
                key: report.bin
                total: 2
                versions:
                  - { "key": "report.bin", "size": 1024, "version_id": "b6ff7d4a", "is_latest": true }
                  - { "key": "report.bin", "size": 512, "version_id": "0a1c33ef" }
        "503":
          description: S3 не настроен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /files/{key}/tags:
    put:
      summary: Заменить теги объекта (закрытый раздел)
      tags: [storage]
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
            example: { "kind": "report", "owner": "evgenza" }
      responses:
        "200":
          description: Теги обновлены
          content:
            application/json:
              example: { "key": "report.bin", "tags": { "kind": "report" } }
        "401":
          description: Нет или недействительный токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gw/messages:
    get:
      summary: Список сообщений через gateway (HTTP+JSON)
      tags: [gateway]
      responses:
        "200":
          description: Проксированный ответ app
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      summary: Создать сообщение через gateway (HTTP+JSON)
      description: |
        Если клиент не передал Idempotency-Key, gateway генерирует его сам,
        поэтому повторы запроса на пути gateway → app не создают дубликатов.
      tags: [gateway]
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: Ключ идемпотентности (по умолчанию генерирует gateway)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: через gateway
      responses:
        "201":
          description: Сообщение сохранено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Нет или недействительный токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gw/grpc/messages:
    get:
      summary: Список сообщений через gRPC server streaming
      description: Gateway читает поток сообщений из app по gRPC и отдаёт их построчно (NDJSON)
      tags: [gateway-grpc]
      responses:
        "200":
          description: Поток сообщений, по одному JSON-объекту на строку
          content:
            application/x-ndjson: {}
    post:
      summary: Создать сообщение через gRPC unary-вызов
      tags: [gateway-grpc]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: через grpc
      responses:
        "201":
          description: Сообщение сохранено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Нет или недействительный токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gw/grpc/messages/batch:
    post:
      summary: Пакетное создание через gRPC client streaming
      description: Gateway отправляет каждый текст отдельным сообщением в клиентский поток, app отвечает одной сводкой
      tags: [gateway-grpc]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [texts]
              properties:
                texts:
                  type: array
                  items:
                    type: string
                  example: [первое, второе, третье]
      responses:
        "201":
          description: Сводка по созданным сообщениям
          content:
            application/json:
              example: { "created": 3, "ids": [101, 102, 103] }
        "401":
          description: Нет или недействительный токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gw/grpc/chat:
    post:
      summary: Эхо-чат через gRPC bidirectional streaming
      description: Каждая реплика уходит в двунаправленный поток, ответы собираются в том же порядке
      tags: [gateway-grpc]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notes]
              properties:
                notes:
                  type: array
                  items:
                    type: string
                  example: [привет, как дела]
      responses:
        "200":
          description: Ответы сервера
          content:
            application/json:
              example: { "replies": ["эхо: привет", "эхо: как дела"] }